chat model v1
The chat model allows to chat based on a QA decision graph. Whenever the agent perceives that it has an answer to the current question, it sends the answer and gets the follow up question. If the user tries to get out of context, the bot mantains it's intent to politely get an answer.